--- lib/models/provider/listItemState.dart
@@ -1,8 +1,0 @@
-class TasksListItemState {
-  bool isOpen;
-
-  TasksListItemState({
-    this.isOpen,
-  });
-  
-}
