+++ lib/utils/convert/epochFromDate.dart
@@ -1,0 +1,3 @@
+  int epochFromDate(DateTime dt) {  
+    return dt.millisecondsSinceEpoch;
+  }
